provendo configurações da npool
