+++ src/app/app.module.ts
@@ -1,11 +1,9 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { GlobalConfigModule } from '@shared/global-config/global-config.module';
-import { OmegleServiceModule } from '@shared/omegle-service/omegle-service.module';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { ChatModule } from './pages/chat/chat.module';
-import { NostrModule } from './shared/nostr/nostr.module';
+import { PoolModule } from './shared/pool/pool.module';
 import { ErrorHandlingModule } from '@shared/error-handling/error-handling.module';
 import { ToastrModule } from 'ngx-toastr';
 import { AsyncModalModule } from '@belomonte/async-modal-ngx';
@@ -18,10 +16,8 @@ import { AsyncModalModule } from '@belomonte/async-modal-ngx';
     ChatModule,
     BrowserModule,
     AppRoutingModule,
-    GlobalConfigModule,
-    OmegleServiceModule,
     AsyncModalModule,
-    NostrModule,
+    PoolModule,
     ToastrModule.forRoot({
       timeOut: 10_000,
       positionClass: 'toast-top-center',
