--- tsconfig.worker.json
@@ -1,15 +1,0 @@
-/* To learn more about Typescript configuration file: https://www.typescriptlang.org/docs/handbook/tsconfig-json.html. */
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./out-tsc/worker",
-    "lib": [
-      "es2018",
-      "webworker"
-    ],
-    "types": []
-  },
-  "include": [
-    "src/**/*.worker.ts"
-  ]
-}
